6.4.0.p1 integrate ovxlib and nnapi addon driver
VIVANTE_GALVIP_Unified_Src_drv_OVXLIB-addon_6.4.0.p1.tgz
VIVANTE_GAL_Unified_Src_drv_NNAdapt-addon_6.4.0.p1.noGPL.org.tgz

+++ ovxlib/src/ops/vsi_nn_op_lstmunit_ovxlib.c
@@ -682,7 +682,7 @@ static vsi_bool op_setup
         if( p->local.multi_batch )
         {
             /* projection FC on NN requires quantized input */
-            attr.dtype.scale = 0.007866097716834601;
+            attr.dtype.scale = (float)0.007866097716834601;
             attr.dtype.zero_point = 128;
             attr.dtype.qnt_type = VSI_NN_QNT_TYPE_AFFINE_ASYMMETRIC;
             attr.dtype.vx_type = VSI_NN_TYPE_UINT8;
